Consolidate fsharp tests into main test assembly
